chore(cli): move version table to a utility
